fix: inventory system review — fix item duplication, panics, and architectural issues
- Fix critical item duplication bug in pickup system (partial pickup now updates dropped item count)
- ItemRegistry::by_name returns Option instead of panicking on unknown items
- Add try_get() with bounds checking to ItemRegistry
- Route items to material_bag by ItemType (Block/Material), with overflow to main_bag
- count_item returns u32 to prevent overflow with 80 slots
- Remove duplicate PickupConfig insertion from ItemPlugin
- Add despawn timer check in dropped_item_physics_system
- Extract spawn_tile_drops helper, eliminating duplicated drop-spawning code
- Gate magnetism/pickup systems behind AppState::InGame
- Fix calculate_drops chance range to inclusive [0.0, 1.0]
- Refactor Hotbar to reference-based (stores item_id, resolves count from Inventory)
- Add active_set field and toggle_set() to Hotbar
- Add Equipment <-> Inventory integration (equip_from_inventory, unequip_to_inventory)
- Extract on_slot_hover/on_slot_unhover observer helpers, removing ~10 duplicated closures
- Add HexColor alpha channel support (#RRGGBBAA)
- Document CraftingPlugin as placeholder with TODO items
- Add 18 new tests (160 total, all passing)

+++ src/crafting/registry.rs
@@ -53,7 +53,7 @@ impl RecipeRegistry {
                 r.unlocked_by.is_unlocked(unlocked)
                     && r.ingredients
                         .iter()
-                        .all(|ing| inventory.count_item(&ing.item_id) >= ing.count)
+                        .all(|ing| inventory.count_item(&ing.item_id) >= ing.count as u32)
             })
             .collect()
     }
